requirements.txt
python-dotenv==1.2.1
streamlit==1.51.0
openai==2.8.1